Object refactoring
