+++ firebird-ng/src/app/app.config.ts
@@ -2,7 +2,7 @@ import {APP_INITIALIZER, ApplicationConfig, importProvidersFrom} from '@angular/
 import { provideRouter } from '@angular/router';
 import { routes } from './app.routes';
 import {provideAnimations} from "@angular/platform-browser/animations";
-import {ServerConfigService} from "./server-config.service";
+import {ServerConfigService} from "./services/server-config.service";
 import {provideHttpClient, withFetch} from "@angular/common/http";
 
 export const appConfig: ApplicationConfig = {
