+++ firebird-ng/src/app/components/resource-select/resource-select.component.scss
