+++ firebird-ng/src/app/data-pipelines/three-event.processor.ts
