--- firebird-ng/src/app/edm4hep-listing/edm4hep-listing.component.html
@@ -1,7 +1,0 @@
-<div class="container mt-3">
-  <h2 class="card-title">Configure geometry pipeline</h2>
-
-  <p>edm4hep-listing works!</p>
-  <p>{{message}}</p>
-
-</div>
