--- firebird-ng/src/app/event-loader.worker.ts
@@ -1,15 +1,0 @@
-/// <reference lib="webworker" />
-
-import {addWarning} from "@angular-devkit/build-angular/src/utils/webpack-diagnostics";
-import {loadJSONFileEvents, loadZipFileEvents} from "./utils/data-fetching.utils";
-
-addEventListener('message', ({ data }) => {
-
-  if(data.endsWith('.zip')) {
-    loadZipFileEvents(data).then(eventsData=>postMessage(eventsData));
-  } else {
-    loadJSONFileEvents(data).then(eventsData=>postMessage(eventsData));
-  }
-});
-
-
