--- firebird-ng/src/app/file-browser/file-browser.component.html
@@ -1,1 +1,0 @@
-<p>file-browser works!</p>
