--- firebird-ng/src/app/file-browser/file-browser.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-file-browser',
-  standalone: true,
-  imports: [],
-  templateUrl: './file-browser.component.html',
-  styleUrl: './file-browser.component.scss'
-})
-export class FileBrowserComponent {
-
-}
