+++ firebird-ng/src/app/pages/input-config/input-config.component.html
@@ -2,7 +2,7 @@
 <div class="container mt-3">
   <h2 class="card-title">Configure geometry pipeline</h2>
 
-  <img src="assets/diagrams/geometry-pipeline.svg" alt="geometry-pipeline">
+  <img src="../../../assets/diagrams/geometry-pipeline.svg" alt="geometry-pipeline">
 
   <form >
