+++ firebird-ng/src/app/pages/input-config/input-config.component.scss
