+++ firebird-ng/src/app/pages/playground/playground.component.scss
