+++ firebird-ng/src/app/services/game-controller.service.ts
