+++ firebird-ng/src/app/services/geometry.service.spec.ts
